LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/static-router-request-destination.https-expected.txt


FAIL Subreosurce load matched with the requestMethod script condition assert_equals: expected (boolean) true but got (undefined) undefined
PASS Subreosurce load not matched with the requestMethod script condition
